Move post_save triggers to the same location
