+++ api/apps.py
@@ -3,3 +3,5 @@ from django.apps import AppConfig
 
 class ApiConfig(AppConfig):
     name = 'api'
+    def ready(self):
+        import api.triggers
