+++ api/models.py
@@ -1,11 +1,8 @@
-import os
 from django.db import models
 from django.conf import settings
-from django.db.models.signals import post_save
 from django.utils import timezone
 from enumfields import EnumIntegerField
 from enumfields import IntEnum
-from .esconnection import ES_CLIENT
 
 
 class DisasterType(models.Model):
@@ -359,28 +356,3 @@ class Profile(models.Model):
 
     def __str__(self):
         return self.user.username
-
-
-# Save a user profile whenever we create a user
-def create_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance)
-    instance.profile.save()
-post_save.connect(create_profile, sender=settings.AUTH_USER_MODEL)
-
-
-def index_es(sender, instance, created, **kwargs):
-    if ES_CLIENT is not None:
-        ES_CLIENT.index(
-            index=instance.es_index(),
-            doc_type='page',
-            id=instance.es_id(),
-            body=instance.indexing(),
-        )
-
-
-# Avoid automatic indexing during bulk imports
-if os.environ.get('BULK_IMPORT') != '1' and ES_CLIENT is not None:
-    post_save.connect(index_es, sender=Event)
-    post_save.connect(index_es, sender=Appeal)
-    post_save.connect(index_es, sender=FieldReport)
